refactor(Sidebar): Organise partial components and types
