+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/SidebarNavItem.tsx
@@ -1,25 +1,12 @@
-import React, { useState, useContext, useRef } from 'react'
+import React, { useContext, useRef, useState } from 'react'
 import { Transition } from 'react-transition-group'
 
 import { SidebarSubNav } from './SidebarSubNav'
-import { Nav, NavItem } from '../../../common/Nav'
 import { SidebarContext } from './context'
 import { Tooltip } from '../../Tooltip'
 import { TRANSITION_TIMEOUT } from './constants'
-import { StyledNavItem } from './partials/StyledNavItem'
-import { StyledNavItemIcon } from './partials/StyledNavItemIcon'
-import { StyledNavItemText } from './partials/StyledNavItemText'
-
-export interface SidebarNavItemProps extends NavItem, Nav<SidebarNavItemProps> {
-  /**
-   * Optiona icon to display to the left of the navigation item.
-   */
-  icon?: React.ReactNode
-  /**
-   * Optional handler invoked when an item is clicked on.
-   */
-  onClick?: (e: React.MouseEvent<HTMLElement>) => void
-}
+import { StyledNavItem, StyledNavItemIcon, StyledNavItemText } from './partials'
+import { SidebarNavItemProps } from './types'
 
 export const SidebarNavItem = ({
   isActive,
