+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/SidebarNotifications.tsx
@@ -3,28 +3,14 @@ import { IconNotifications } from '@royalnavy/icon-library'
 import { Transition } from 'react-transition-group'
 
 import { SidebarContext } from './context'
-import { ComponentWithClass } from '../../../common/ComponentWithClass'
-import { NotificationsProps } from '../NotificationPanel'
 import { TRANSITION_TIMEOUT } from './constants'
-import { StyledNotificationsSheet } from './partials/StyledNotificationsSheet'
-import { StyledNotificationsSheetButton } from './partials/StyledNotificationsSheetButton'
-import { StyledNotificationDot } from './partials/StyledNotificationDot'
-import { StyledNotificationsLabel } from './partials/StyledNotificationsLabel'
-
-export interface SidebarNotificationsProps extends ComponentWithClass {
-  /**
-   * Toggle whether there are unread notifications.
-   */
-  hasUnreadNotification?: boolean
-  /**
-   * Whether the sheet is initially open.
-   */
-  initialIsOpen?: boolean
-  /**
-   * Collection of Notification item components.
-   */
-  notifications?: React.ReactElement<NotificationsProps>
-}
+import {
+  StyledNotificationDot,
+  StyledNotificationsLabel,
+  StyledNotificationsSheet,
+  StyledNotificationsSheetButton,
+} from './partials'
+import { SidebarNotificationsProps } from './types'
 
 export const SidebarNotifications = ({
   initialIsOpen,
