+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/SidebarUserItem.tsx
@@ -1,16 +1,10 @@
 import React from 'react'
-
-import { NavItem } from '../../../common/Nav'
-import { StyledUserItem } from './partials/StyledUserItem'
-import { StyledUserItemIcon } from './partials/StyledUserItemIcon'
-import { StyledUserItemText } from './partials/StyledUserItemText'
-
-export interface SidebarUserItemProps extends NavItem {
-  /**
-   * Optional icon to display to the left of the user navigation item.
-   */
-  icon?: React.ReactNode
-}
+import {
+  StyledUserItem,
+  StyledUserItemIcon,
+  StyledUserItemText,
+} from './partials'
+import { SidebarUserItemProps } from './types'
 
 export const SidebarUserItem = ({ icon, link }: SidebarUserItemProps) => {
   const linkElement = link as React.ReactElement
