+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/index.ts
@@ -4,3 +4,4 @@ export * from './SidebarNavItem'
 export * from './SidebarWrapper'
 export * from './SidebarUser'
 export * from './SidebarUserItem'
+export * from './types'
