+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/partials/index.ts
@@ -1,0 +1,29 @@
+export { StyledHead } from './StyledHead'
+export { StyledHeadIcon } from './StyledHeadIcon'
+export { StyledHeadTitle } from './StyledHeadTitle'
+export { StyledSidebar } from './StyledSidebar'
+
+export { StyledUserAvatar } from './StyledUserAvatar'
+export { StyledUserSheetButton } from './StyledUserSheetButton'
+export { StyledUser } from './StyledUser'
+export { StyledSheetList } from './StyledSheetList'
+export { StyledUserText } from './StyledUserText'
+
+export { StyledHandle } from './StyledHandle'
+
+export { StyledNav } from './StyledNav'
+
+export { StyledNavItem } from './StyledNavItem'
+export { StyledNavItemIcon } from './StyledNavItemIcon'
+export { StyledNavItemText } from './StyledNavItemText'
+
+export { StyledNotificationsSheet } from './StyledNotificationsSheet'
+export { StyledNotificationsSheetButton } from './StyledNotificationsSheetButton'
+export { StyledNotificationDot } from './StyledNotificationDot'
+export { StyledNotificationsLabel } from './StyledNotificationsLabel'
+
+export { StyledUserItem } from './StyledUserItem'
+export { StyledUserItemIcon } from './StyledUserItemIcon'
+export { StyledUserItemText } from './StyledUserItemText'
+
+export { StyledWrapper } from './StyledWrapper'
