+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/types/index.ts
@@ -1,0 +1,6 @@
+export type { SidebarProps } from './SidebarProps'
+export type { SidebarUserProps } from './SidebarUserProps'
+export type { SidebarNavProps } from './SidebarNavProps'
+export type { SidebarNavItemProps } from './SidebarNavItemProps'
+export type { SidebarNotificationsProps } from './SidebarNotificationsProps'
+export type { SidebarUserItemProps } from './SidebarUserItemProps'
